LunaScript: Fixed BMP with transparent key color
Added an ability to set any color as a transparent key color which is required for backward compatibility with old LunaDLL sprites

One of bugs at #270

+++ src/script/luna/autocode.cpp
@@ -1151,7 +1151,7 @@ void Autocode::Do(bool init)
         {
             // Only allow loading image during init phase
             if(init)
-                Renderer::Get().LoadBitmapResource(MyString, (int)Target, (int)Param1);
+                Renderer::Get().LoadBitmapResource(MyString, (int)Target, (uint32_t)(int32_t)Param1);
             expire();
             break;
         }
